refactor: [user] Response(controller), Request(controller) (domain) 분리
