+++ src/main/java/site/travellaboratory/be/user/domain/_auth/request/LoginRequest.java
@@ -1,0 +1,8 @@
+package site.travellaboratory.be.user.domain._auth.request;
+
+public record LoginRequest(
+    String username,
+    String password
+) {
+}
+
