+++ src/main/java/site/travellaboratory/be/user/domain/_auth/request/UserNicknameRequest.java
@@ -1,0 +1,6 @@
+package site.travellaboratory.be.user.domain._auth.request;
+
+public record UserNicknameRequest(
+    String nickname
+) {
+}
