+++ src/main/java/site/travellaboratory/be/user/domain/_auth/request/UsernameRequest.java
@@ -1,0 +1,7 @@
+package site.travellaboratory.be.user.domain._auth.request;
+
+public record UsernameRequest(
+    String username
+) {
+
+}
