+++ src/main/java/site/travellaboratory/be/user/presentation/_auth/controller/PwInquiryController.java
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.travellaboratory.be.user.application._auth.PwInquiryService;
-import site.travellaboratory.be.user.presentation._auth.response.pwinquiry.PwInquiryEmailRequest;
+import site.travellaboratory.be.user.presentation._auth.request.PwInquiryEmailRequest;
 import site.travellaboratory.be.user.presentation._auth.response.pwinquiry.PwInquiryEmailResponse;
-import site.travellaboratory.be.user.presentation._auth.response.pwinquiry.PwInquiryRenewalRequest;
-import site.travellaboratory.be.user.presentation._auth.response.pwinquiry.PwInquiryVerificationRequest;
+import site.travellaboratory.be.user.presentation._auth.request.PwInquiryRenewalRequest;
+import site.travellaboratory.be.user.presentation._auth.request.PwInquiryVerificationRequest;
 import site.travellaboratory.be.user.presentation._auth.response.pwinquiry.PwInquiryVerificationResponse;
 
 @RestController
