+++ src/main/java/site/travellaboratory/be/user/presentation/_auth/request/PwInquiryEmailRequest.java
@@ -1,0 +1,6 @@
+package site.travellaboratory.be.user.presentation._auth.request;
+
+public record PwInquiryEmailRequest(
+    String username
+) {
+}
