--- src/main/java/site/travellaboratory/be/user/presentation/_auth/response/pwinquiry/PwInquiryEmailRequest.java
@@ -1,6 +1,0 @@
-package site.travellaboratory.be.user.presentation._auth.response.pwinquiry;
-
-public record PwInquiryEmailRequest(
-    String username
-) {
-}
