--- src/main/java/site/travellaboratory/be/user/presentation/_auth/response/userauthentication/LoginRequest.java
@@ -1,8 +1,0 @@
-package site.travellaboratory.be.user.presentation._auth.response.userauthentication;
-
-public record LoginRequest(
-    String username,
-    String password
-) {
-}
-
