--- src/main/java/site/travellaboratory/be/user/presentation/_auth/response/userverification/UserNicknameRequest.java
@@ -1,6 +1,0 @@
-package site.travellaboratory.be.user.presentation._auth.response.userverification;
-
-public record UserNicknameRequest(
-    String nickname
-) {
-}
