--- src/main/java/site/travellaboratory/be/user/presentation/_auth/response/userverification/UsernameRequest.java
@@ -1,7 +1,0 @@
-package site.travellaboratory.be.user.presentation._auth.response.userverification;
-
-public record UsernameRequest(
-    String username
-) {
-
-}
