+++ src/main/java/site/travellaboratory/be/user/presentation/controller/UserProfileWriterController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import site.travellaboratory.be.user.application.UserProfileWriterService;
 import site.travellaboratory.be.common.annotation.UserId;
-import site.travellaboratory.be.user.presentation.response.writer.UserProfileUpdateRequest;
+import site.travellaboratory.be.user.domain.request.UserProfileUpdateRequest;
 import site.travellaboratory.be.user.presentation.response.writer.UserProfileUpdateResponse;
 
 @RestController
